10/10/2021:
-Huge changes to Collision Model and its implementations;
-Switched calls of idMath::Fabs to Math::abs.

+++ src/main/java/neo/CM/CollisionModel_contents.java
@@ -1,17 +1,13 @@
 package neo.CM;
 
-import neo.CM.CollisionModel_local.cm_edge_s;
-import neo.CM.CollisionModel_local.cm_vertex_s;
+import neo.CM.AbstractCollisionModel_local.cm_edge_s;
+import neo.CM.AbstractCollisionModel_local.cm_vertex_s;
 import neo.idlib.math.Plane.idPlane;
 import neo.idlib.math.Pluecker.idPluecker;
 
 import static neo.idlib.math.Math_h.FLOATSIGNBITSET;
 
-/**
- *
- */
 public class CollisionModel_contents {
-
     /*
      ===============================================================================
 
@@ -25,11 +21,11 @@ public class CollisionModel_contents {
      ================
      */
     static void CM_SetTrmEdgeSidedness(cm_edge_s edge, final idPluecker bpl, final idPluecker epl, final int bitNum) {
-        if (0 == (edge.sideSet & (1 << bitNum))) {
+        if (0 == (edge.sideSet & (1L << bitNum))) {
             float fl;
             fl = bpl.PermutedInnerProduct(epl);
-            edge.side = (edge.side & ~(1 << bitNum)) | (FLOATSIGNBITSET(fl) << bitNum);
-            edge.sideSet |= (1 << bitNum);
+            edge.side = (edge.side & ~(1L << bitNum)) | ((long) FLOATSIGNBITSET(fl) << bitNum);
+            edge.sideSet |= (1L << bitNum);
         }
     }
 
@@ -39,16 +35,18 @@ public class CollisionModel_contents {
      ================
      */
     static void CM_SetTrmPolygonSidedness(cm_vertex_s v, final idPlane plane, final int bitNum) {
-        if (0 == (v.sideSet & (1 << bitNum))) {
+        if (0 == (v.sideSet & (1L << bitNum))) {
             float fl;
             fl = plane.Distance(v.p);
             /* cannot use float sign bit because it is undetermined when fl == 0.0f */
             if (fl < 0.0f) {
-                v.side |= (1 << bitNum);
+                v.side |= (1L << bitNum);
             } else {
-                v.side &= ~(1 << bitNum);
+                v.side &= ~(1L << bitNum);
             }
-            v.sideSet |= (1 << bitNum);
+            v.sideSet |= (1L << bitNum);
         }
     }
+
+
 }
